Support clients handling the data passing internally
